restructured and added slice for problem container
